pyproject.toml
[tool.poetry]
name = "matemux"
version = "0.1.3"
description = "tmux session generator"
authors = ["Kamyab Taghizadeh <kamyab.zad@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/kamyabzad/matemux"
repository = "https://github.com/kamyabzad/matemux"
keywords = ["tmux", "console", "automation"]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
matemux = "matemux.__main__:main"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
